refactor(hr): okr/v2 KeyResult + Alignment 组改引用 common::models（4 叶删 inline + 3 跨叶 consumer 改路径）(#336)
